+++ davarocr/mmcv/utils/__init__.py
@@ -2,18 +2,12 @@
 ##################################################################################################
 # Copyright Info :    Copyright (c) Davar Lab @ Hikvision Research Institute. All rights reserved.
 # Filename       :    __init__.py
-# Abstract       :   Some custommized implementations that different from official mmcv.
+# Abstract       :   progressbar init file.
 
 # Current Version:    1.0.0
-# Date           :    2021-05-01
+# Date           :    2022-05-18
 ##################################################################################################
 """
 
-from .runner import DavarCheckpointHook, ReduceonplateauLrUpdaterHook
-from .utils import DavarProgressBar
 
-__all__ = [
-    'DavarCheckpointHook',
-    'ReduceonplateauLrUpdaterHook',
-    'DavarProgressBar'
-]
+from .davar_progressbar import DavarProgressBar